premier commit du projet laravel-social

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Post;
 use Illuminate\Http\Request;
 use App\Http\Requests;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
@@ -27,18 +28,21 @@ class PostController extends Controller
 
     public function getDashboard()
     {
-        $posts = Post::all();
+        $posts = Post::orderBy('created_at','desc')->get();
         return view('dashboard',['posts'=>$posts]);
     }
 
     public function getSupprimerPost($post_id)
     {
         $post=Post::where('id',$post_id)->first();
+        if (Auth::user() != $post->user ){
+            return redirect()->back();
+        }
         $post->delete();
         return redirect()->route('dashboard')->with(['message'=>'Message supprimé avec succes.']);
     }
 
 
 
-    
+
 }
